+++ x/metadata/keeper/genesis.go
@@ -8,6 +8,7 @@ import (
 
 // InitGenesis creates the initial genesis state for the metadata module.
 func (k Keeper) InitGenesis(ctx sdk.Context, data *types.GenesisState) {
+	k.SetParams(ctx, data.Params)
 	k.SetOSLocatorParams(ctx, data.OSLocatorParams)
 	if err := data.Validate(); err != nil {
 		panic(err)