+++ x/metadata/keeper/params.go
@@ -8,11 +8,8 @@ import (
 
 // GetParams returns the total set of metadata parameters.
 func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
-	k.paramSpace.GetParamSet(ctx, &params)
-	return params
+	return types.Params{} // there are currently no params so no further action required here.
 }
 
 // SetParams sets the metadata parameters to the param space.
-func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
-	k.paramSpace.SetParamSet(ctx, &params)
-}
+func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {} // currently no params are supported